refactor(api): split backup.go, discovery.go, and configsync.go into cohesive files
Pure relocation, no behavior or signature change. Splits the three largest
non-test source files in internal/api along their existing section
boundaries into same-package sibling files:

- backup.go (1264) -> backup.go (core + CRUD) + backup_restore.go
  + backup_retention.go + backup_scheduler.go
- discovery.go (1231) -> discovery.go (handlers) + discovery_diff.go
  + discovery_confirm.go
- configsync.go (1359) -> configsync.go (core + envelope types + helpers)
  + configsync_export.go + configsync_import.go + configsync_apply.go

No file now exceeds ~680 lines. The set of top-level declarations is
byte-identical before and after the split; only per-file import lists
changed (goimports-pruned). Verified: go build, go vet, golangci-lint
(0 issues), and the full internal/api test suite all pass.
